Refactoring app and started database config
