--- data.json
@@ -1,25 +1,0 @@
-{
-  "students": [],
-  "teachers": [
-    {
-      "id": 1,
-      "avatar_url": "https://www.nea.org/sites/default/files/legacy/2020/04/new_teacher.jpeg",
-      "name": "Ericka Marshall",
-      "birth": 597888000000,
-      "degree": "doctoral",
-      "class_type": "Online",
-      "subjects": "Math, English",
-      "created_at": 1601388557856
-    },
-    {
-      "id": 2,
-      "avatar_url": "https://images.squarespace-cdn.com/content/v1/573e57871bbee0d6dea60fff/1551203818326-Y2YY9W2OHZT2R28UZ2UC/ke17ZwdGBToddI8pDm48kG87Sfbgg29A4BYEDq3OXvgUqsxRUqqbr1mOJYKfIPR7LoDQ9mXPOjoJoqy81S2I8N_N4V1vUb5AoIIIbLZhVYxCRW4BPu10St3TBAUQYVKcf4OxbJOyh_wHUnyc4kQLQ6SBshRGOku7c30Y_IRDNPta8R2IY5BHMaEj1zOWoDTZ/what-is-teacher-burnout.jpg",
-      "name": "Martin Austin",
-      "birth": 263174400000,
-      "degree": "master",
-      "class_type": "Traditional",
-      "subjects": "Math, Programming",
-      "created_at": 1601661597113
-    }
-  ]
-}
