+++ src/app/views/deleteBtn.njk
