+++ src/app/views/teachers/show.njk
