+++ src/routes.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const routes = express.Router()
-const teachers = require('./controllers/teachers')
-const students = require('./controllers/students')
+const teachers = require('./app/controllers/teachers')
+const students = require('./app/controllers/students')
 
 routes.get('/', (req, res) => {
     return res.redirect("/teachers")
